// src/types/BlockComponentType.ts
enum BlockComponentType {
    HeaderBlock = 'HeaderBlock',
    TextBlock = 'TextBlock',
    LinkBlock = 'LinkBlock',
    ImageBlock = 'ImageBlock',
    ConventionBlock = 'ConventionBlock',
    ConvetionScheduleBlock = 'ConvetionScheduleBlock',
    ConventionEventBlock = 'ConventionEventBlock',
    ArrangementBlock = 'ArrangementBlock',
    ArrangementInfoBlock = 'ArrangementInfoBlock',
}

export default BlockComponentType;